Create group R transaction type
