+++ front-end/src/app/shared/models/transaction-groups/transaction-group-m.model.ts
@@ -17,16 +17,26 @@ export class TransactionGroupM extends TransactionGroup {
       templateMap.election_other_description,
       templateMap.date,
       templateMap.amount,
-      templateMap.aggregate,
       templateMap.purpose_description,
       templateMap.category_code,
+      templateMap.committee_fec_id,
+      templateMap.committee_name,
+      templateMap.candidate_fec_id,
+      templateMap.candidate_last_name,
+      templateMap.candidate_first_name,
+      templateMap.candidate_middle_name,
+      templateMap.candidate_prefix,
+      templateMap.candidate_suffix,
+      templateMap.candidate_office,
+      templateMap.candidate_state,
+      templateMap.candidate_district,
       templateMap.memo_code,
       templateMap.text4000,
     ].filter((val) => !!val);
   }
 
   getContactTypeOptions(): PrimeOptions {
-    return LabelUtils.getPrimeOptions(ContactTypeLabels, [ContactTypes.ORGANIZATION]);
+    return LabelUtils.getPrimeOptions(ContactTypeLabels, [ContactTypes.COMMITTEE]);
   }
 
   hasEmployerInput(): boolean {
@@ -34,7 +44,7 @@ export class TransactionGroupM extends TransactionGroup {
   }
 
   hasCommitteeFecIdInput(): boolean {
-    return false;
+    return true;
   }
 
   hasElectionInformationInput(): boolean {
@@ -42,6 +52,6 @@ export class TransactionGroupM extends TransactionGroup {
   }
 
   hasCandidateInformationInput(): boolean {
-    return false;
+    return true;
   }
 }
