+++ front-end/src/app/shared/models/transaction-groups/transaction-group-r.model.spec.ts
@@ -2,18 +2,18 @@ import { TestBed } from '@angular/core/testing';
 import { LabelUtils } from 'app/shared/utils/label.utils';
 import { testTemplateMap } from 'app/shared/utils/unit-test.utils';
 import { ContactTypeLabels, ContactTypes } from '../contact.model';
-import { TransactionGroupL } from './transaction-group-l.model';
+import { TransactionGroupR } from './transaction-group-r.model';
 
-describe('TransactionGroupL', () => {
-  let component: TransactionGroupL;
+describe('TransactionGroupR', () => {
+  let component: TransactionGroupR;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [],
-      providers: [TransactionGroupL],
+      providers: [TransactionGroupR],
     });
 
-    component = TestBed.inject(TransactionGroupL);
+    component = TestBed.inject(TransactionGroupR);
   });
 
   it('should create', () => {
@@ -29,7 +29,7 @@ describe('TransactionGroupL', () => {
   });
 
   it('#getContactTypeOptions happy path', () => {
-    const expectedRetval = LabelUtils.getPrimeOptions(ContactTypeLabels, [ContactTypes.COMMITTEE]);
+    const expectedRetval = LabelUtils.getPrimeOptions(ContactTypeLabels, [ContactTypes.ORGANIZATION]);
     const retval = component.getContactTypeOptions();
     expect(JSON.stringify(expectedRetval) === JSON.stringify(retval)).toBeTruthy();
   });
@@ -41,7 +41,7 @@ describe('TransactionGroupL', () => {
 
   it('#hasCommitteeFecIdInput happy path', () => {
     const retval = component.hasCommitteeFecIdInput();
-    expect(retval).toBeTrue();
+    expect(retval).toBeFalse();
   });
 
   it('#hasElectionInformationInput happy path', () => {
@@ -51,6 +51,6 @@ describe('TransactionGroupL', () => {
 
   it('#hasCandidateInformationInput happy path', () => {
     const retval = component.hasCandidateInformationInput();
-    expect(retval).toBeTrue();
+    expect(retval).toBeFalse();
   });
 });
