+++ front-end/src/app/shared/models/transaction-groups/transaction-group-r.model.ts
@@ -3,7 +3,7 @@ import { ContactTypeLabels, ContactTypes } from '../contact.model';
 import { TransactionTemplateMapType } from '../transaction-type.model';
 import { TransactionGroup } from './transaction-group.model';
 
-export class TransactionGroupL extends TransactionGroup {
+export class TransactionGroupR extends TransactionGroup {
   getFormProperties(templateMap: TransactionTemplateMapType): string[] {
     return [
       'entity_type',
@@ -17,26 +17,16 @@ export class TransactionGroupL extends TransactionGroup {
       templateMap.election_other_description,
       templateMap.date,
       templateMap.amount,
+      templateMap.aggregate,
       templateMap.purpose_description,
       templateMap.category_code,
-      templateMap.committee_fec_id,
-      templateMap.committee_name,
-      templateMap.candidate_fec_id,
-      templateMap.candidate_last_name,
-      templateMap.candidate_first_name,
-      templateMap.candidate_middle_name,
-      templateMap.candidate_prefix,
-      templateMap.candidate_suffix,
-      templateMap.candidate_office,
-      templateMap.candidate_state,
-      templateMap.candidate_district,
       templateMap.memo_code,
       templateMap.text4000,
     ].filter((val) => !!val);
   }
 
   getContactTypeOptions(): PrimeOptions {
-    return LabelUtils.getPrimeOptions(ContactTypeLabels, [ContactTypes.COMMITTEE]);
+    return LabelUtils.getPrimeOptions(ContactTypeLabels, [ContactTypes.ORGANIZATION]);
   }
 
   hasEmployerInput(): boolean {
@@ -44,7 +34,7 @@ export class TransactionGroupL extends TransactionGroup {
   }
 
   hasCommitteeFecIdInput(): boolean {
-    return true;
+    return false;
   }
 
   hasElectionInformationInput(): boolean {
@@ -52,6 +42,6 @@ export class TransactionGroupL extends TransactionGroup {
   }
 
   hasCandidateInformationInput(): boolean {
-    return true;
+    return false;
   }
 }
