+++ front-end/src/app/shared/models/transaction-types/CONTRIBUTION_TO_CANDIDATE.model.ts
@@ -2,11 +2,11 @@ import { LabelUtils } from 'app/shared/utils/label.utils';
 import { schema } from 'fecfile-validate/fecfile_validate_js/dist/CANDIDATE_CONTRIBUTIONS';
 import { SchBTransactionType } from '../schb-transaction-type.model';
 import { SchBTransaction, ScheduleBTransactionTypeLabels, ScheduleBTransactionTypes } from '../schb-transaction.model';
-import { TransactionGroupL } from '../transaction-groups/transaction-group-l.model';
+import { TransactionGroupM } from '../transaction-groups/transaction-group-m.model';
 import { STANDARD_CONTROLS, TransactionNavigationControls } from '../transaction-navigation-controls.model';
 
 export class CONTRIBUTION_TO_CANDIDATE extends SchBTransactionType {
-  transactionGroup = new TransactionGroupL();
+  transactionGroup = new TransactionGroupM();
   title = LabelUtils.get(ScheduleBTransactionTypeLabels, ScheduleBTransactionTypes.CONTRIBUTION_TO_CANDIDATE);
   schema = schema;
   override showAggregate = false;
