+++ front-end/src/app/shared/models/transaction-types/CONTRIBUTION_TO_CANDIDATE_VOID.model.ts
@@ -2,12 +2,12 @@ import { LabelUtils } from 'app/shared/utils/label.utils';
 import { schema } from 'fecfile-validate/fecfile_validate_js/dist/CANDIDATE_CONTRIBUTIONS';
 import { SchBTransactionType } from '../schb-transaction-type.model';
 import { SchBTransaction, ScheduleBTransactionTypeLabels, ScheduleBTransactionTypes } from '../schb-transaction.model';
-import { TransactionGroupL } from '../transaction-groups/transaction-group-l.model';
+import { TransactionGroupM } from '../transaction-groups/transaction-group-m.model';
 import { STANDARD_CONTROLS, TransactionNavigationControls } from '../transaction-navigation-controls.model';
 import { PurposeDescriptionLabelSuffix } from '../transaction-type.model';
 
 export class CONTRIBUTION_TO_CANDIDATE_VOID extends SchBTransactionType {
-  transactionGroup = new TransactionGroupL();
+  transactionGroup = new TransactionGroupM();
   title = LabelUtils.get(ScheduleBTransactionTypeLabels, ScheduleBTransactionTypes.CONTRIBUTION_TO_CANDIDATE_VOID);
   schema = schema;
   override negativeAmountValueOnly = true;
